+++ src/main/scala/draco/DerivationResolvable.scala
@@ -8,7 +8,7 @@ object DerivationResolvable extends App with DracoType {
   override lazy val typeDefinition: TypeDefinition = TypeLoader.loadType(TypeName ("DerivationResolvable", _namePackage = Seq ("draco")))
   lazy val dracoType: Type[DerivationResolvable] = Type[DerivationResolvable] (typeDefinition)
   lazy val domainType: Domain[Draco] = Domain[Draco] (typeDefinition)
-  def w0(m: TypeDefinition): Boolean = m.dracoAspect.derivation.exists(anc => anc.namePackage.headOption.contains("draco") && { val a = draco.TypeLoader.loadType(anc); draco.DracoAspect.isEmpty(a.dracoAspect) && draco.DomainAspect.isEmpty(a.domainAspect) && draco.RuleAspect.isEmpty(a.ruleAspect) && draco.ActorAspect.isEmpty(a.actorAspect) })
+  def w0(m: TypeDefinition): Boolean = draco.DracoAspect.parents(m.dracoAspect).exists(anc => anc.namePackage.headOption.contains("draco") && { val a = draco.TypeLoader.loadType(anc); draco.DracoAspect.isEmpty(a.dracoAspect) && draco.DomainAspect.isEmpty(a.domainAspect) && draco.RuleAspect.isEmpty(a.ruleAspect) && draco.ActorAspect.isEmpty(a.actorAspect) })
   private lazy val action: RhsContext => Unit = (ctx: RhsContext) => {
       val m: TypeDefinition = ctx.get[TypeDefinition]("$m")
       ctx.insert(Problem(m.typeName, s"member ${m.typeName.name} derives from a draco type that does not resolve to a definition"))